[WIP] Some more reorganization
Generalized `BeforeContextAccumulator` and renamed it to `SlidingAccumulartor` (yes, again).
