+++ src/core/results_collection/top_bracket.rs
