+++ src/core/sliding_accumulator.rs
@@ -1,16 +1,16 @@
 use std::collections::VecDeque;
 
-/// A FIFO-like "before" context accumulator: when at capacity every new line that is fed
-/// will pop the oldest line stored in the accumulator.
+/// A FIFO accumulator: when at capacity every new item that is fed
+/// will pop the oldest item stored in the accumulator.
 ///
 #[derive(Debug)]
-pub struct BeforeContextAccumulator {
-    data: Option<VecDeque<String>>,
+pub struct SlidingAccumulator<T> {
+    data: Option<VecDeque<T>>,
 }
 
-impl BeforeContextAccumulator {
+impl<T> SlidingAccumulator<T> {
     /// Creates a new [`SlidingAccumulator`] with capacity `capacity`.
-    /// `capacity` can be 0, in which case [`feed`] does nothing.
+    /// `capacity` can be 0, in which case [`SlidingAccumulator::feed`] does nothing.
     ///
     pub fn new(capacity: usize) -> Self {
         Self {
@@ -18,24 +18,26 @@ impl BeforeContextAccumulator {
         }
     }
 
-    /// Pushes a line into the accumulator.
-    /// If the accumulator is at capacity, the oldest stored line is popped.
+    /// Pushes an item into the accumulator.
+    /// If the accumulator is at capacity, the oldest stored item is popped.
     /// If the capacity is zero does nothing.
     ///
-    pub fn feed(&mut self, line: String) {
+    pub fn feed(&mut self, item: T) {
         if let Some(data) = self.data.as_mut() {
             if data.len() == data.capacity() {
                 data.pop_front();
             }
 
-            data.push_back(line);
+            data.push_back(item);
         }
     }
+}
 
-    /// Returns the accumulated lines as a [`Vec<String>`].
+impl<T: Clone> SlidingAccumulator<T> {
+    /// Returns the accumulated items as a [`Vec<T>`].
     ///
-    pub fn snapshot(&self) -> Option<Vec<String>> {
-        self.data.as_ref().map(|data| Vec::from(data.clone()))
+    pub fn snapshot(&self) -> Option<Vec<T>> {
+        self.data.as_ref().map(|data| Vec::from((*data).clone()))
     }
 }
 
@@ -45,21 +47,21 @@ mod test {
 
     #[test]
     fn sliding_accumulator_constructor_zero_capacity() {
-        let acc = BeforeContextAccumulator::new(0);
+        let acc = SlidingAccumulator::<String>::new(0);
         assert_eq!(acc.data, None);
     }
 
     #[test]
     fn sliding_accumulator_constructor() {
-        let acc = BeforeContextAccumulator::new(3);
-        assert_eq!(acc.data, Some(VecDeque::new()));
+        let acc = SlidingAccumulator::<String>::new(3);
+        assert!(acc.data.as_ref().unwrap().is_empty());
         assert_eq!(acc.data.unwrap().capacity(), 3);
     }
 
     #[test]
     fn sliding_accumulator_feed() {
-        let mut acc = BeforeContextAccumulator::new(3);
-        assert_eq!(acc.data, Some(VecDeque::from([])));
+        let mut acc = SlidingAccumulator::new(3);
+        assert!(acc.data.as_ref().unwrap().is_empty());
         acc.feed(String::from("one"));
         assert_eq!(acc.data, Some(VecDeque::from([String::from("one")])));
         acc.feed(String::from("two"));
@@ -89,7 +91,7 @@ mod test {
 
     #[test]
     fn sliding_accumulator_feed_zero_capacity() {
-        let mut acc = BeforeContextAccumulator::new(0);
+        let mut acc = SlidingAccumulator::new(0);
         assert_eq!(acc.data, None);
         acc.feed(String::from("something"));
         assert_eq!(acc.data, None);
@@ -97,7 +99,7 @@ mod test {
 
     #[test]
     fn sliding_accumulator_snapshot() {
-        let mut acc = BeforeContextAccumulator::new(3);
+        let mut acc = SlidingAccumulator::new(3);
         assert_eq!(acc.snapshot(), None);
         acc.feed(String::from("one"));
         assert_eq!(acc.snapshot(), Some(vec![String::from("one")]));
